# nakar-shared/openapi.yaml
openapi: 3.0.0
info:
  title: NAKAR
  version: 1.0.0
servers:
  - url: http://localhost:1338
    description: Development
  - url: http://nakar.mni.thm.de:1338
    description: Production
...
paths:
...
  /note:
    post:
      tags:
        - Notes
      operationId: postNote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nodeIds:
                  type: array
                  items:
                    type: string
                content:
                  type: string
                canvasId:
                  type: string
              required:
                - nodeIds
                - content
                - canvasId
      responses:
        "200":
          description: "OK"
  /note/{id}:
    delete:
      tags:
        - Notes
      operationId: deleteNote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "OK"
    put:
      tags:
        - Notes
      operationId: putNote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required:
                - content
      responses:
        "200":
          description: "OK"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...